Beautifulsoup work-around, and results list sort
When searching using 'nondbshow', sometimes the last tr gets mangled.
All that can be extracted is the torrent name and magnet url.  This
seems to be a beautifulsoup bug.

Also, the results of the two searches (SxxExx and XxXX) are
concatenated together.  They were originaly not sorted but now they
are sorted by seed count.