Fix a chronic IOError on OSX when asking the user for feedback
The bug is in ConsoleInput.py.  This is code that is beyond my
capabilities, it is from a stack overflow post.
http://stackoverflow.com/questions/510357/python-read-a-single-character-from-the-user

As it was posted it worked on linux, but on OSX there would be a
periodical bug; approximately every 10th time that the user was asked
there would be an IOError.

The fix that I got from:
http://trac.edgewall.org/ticket/2066#comment:1
seems to fix the problem.  We'll see, it has not been tested on linux
so further research might be needed.

For future reference, I used this Expect script to test:

#!/usr/bin/expect -f

for {set i 1} {$i < 120} {incr i 1} {
  spawn tv nondbshow 'x'
  expect {
      "IOError:" {exit}
      "\nChoice: " {puts "q"}
  }
  send_user "$i\n"
}