Add debugging options to transmission_done.bash
If a copy to destination dir didn't work properly the changes now will
make it easier to find out why.

Each time the bash file is called by transmission, in addition to the
usual info written to the log file, a debug command is written as well.
If that command is copied and then used on the command line, it can be
used to debug what the problem was.