Remove experimental code from Tracking.py
Remove experimental code from Tracking.py and add it to
TrackingTesting.py and have git ignore this file.  It was changing
a lot and creating commit noise.  The code here is to experiment
with generating various reports about the files downloaded.  This
ability might be integrated formally into the main app eventually.

Also rewrote the help string in Search.py to be more descriptive.