Add a VPN test
If not connected to a VPN, alert the user and halt the programme.

A config option is added to config.ini where the user can specify
their own ip address.  Then the prog checks what the internet thinks
is their ip address (using api.ipify.org) and if it matches, stop.